fix next js build erros using suspense and client components
